requirements.txt
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
mysql-connector==2.2.9
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
PyYAML==5.4.1
queuelib==1.5.0
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
Twisted==21.2.0
w3lib==1.22.0
zope.interface==5.3.0
